Update scripts inventory and normalization matrix
Regenerate the scripts inventory manifest (updated generated_at, reference counts and many reference entries/line offsets) to reflect discovery updates across CI/docs/scripts references. Adjust the pipeline normalization field matrix CSV: replace several mapping-status fields' normalization profile from text-based handlers to a lambda-driven strict_enum (canonicalize mapping-status to lowercase mapped/unmapped/missing), and refine the assay-parameter type description to clarify controlled-vocabulary handling while preserving unknown provider lexemes.

+++ src/bioetl/infrastructure/schemas/silver_chembl_extended.py
@@ -46,7 +46,7 @@ CHEMBL_MOLECULE_SCHEMA = pa.schema(
         pa.field("hierarchy_parent_chembl_id", pa.string()),
         pa.field("inchi_key", pa.string()),
         pa.field("inorganic_flag", pa.int64()),
-        pa.field("max_phase", pa.int64()),
+        pa.field("max_phase", pa.float64()),
         pa.field("molecule_id", pa.string(), nullable=False),
         pa.field("molecule_hierarchy", pa.string()),
         pa.field("molecule_properties", pa.string()),
